// UI/src/app/reduxFlow/types.ts
enum TYPES {
    FETCH_MOVIE_DETAILS = '[FETCH MOVIE DETAILS] Fetch',
    FETCH_MOVIE_DETAILS_SUCCESS = '[FETCH MOVIE DETAILS] Success',
    FETCH_MOVIE_DETAILS_FAILURE = '[FETCH MOVIE DETAILS] Failure',

    FETCH_THEATRE_DETAILS = '[FETCH THEATRE DETAILS] Fetch',
    FETCH_THEATRE_DETAILS_SUCCESS = '[FETCH THEATRE DETAILS] Success',
    FETCH_THEATRE_DETAILS_FAILURE = '[FETCH THEATRE DETAILS] Failure',

    PAYMENT_MODAL_HANDLE = '[PAYMENT_MODAL_HANDLE] Payment Modal',

    PAYMENT_MODAL_CONTENT = '[PAYMENT_MODAL_CONTENT] Payment Modal Content',

    SAVE_BOOKING_MOVIE_DETAILS = '[SAVE_BOOKING_MOVIE_DETAILS] Save booking movie details',

    FETCH_SELECTED_MOVIE_DETAILS = '[FETCH_SELECTED_MOVIE_DETAILS] Save booking movie details',
    FETCH_SELECTED_MOVIE_DETAILS_SUCCESS = '[FETCH_SELECTED_MOVIE_DETAILS_SUCCESS] Save booking movie details',
    FETCH_SELECTED_MOVIE_DETAILS_FAILUREE = '[FETCH_SELECTED_MOVIE_DETAILS] Save booking movie details',

    SAVE_BOOKING_DETAILS_TO_DB = '[SAVE_BOOKING_DETAILS_TO_DB] Save booking details in DB'
}

export {
    TYPES
}